added getting data from db and getting data from response by swiperefresh

+++ app/src/main/java/kireev/ftshw/project/Database/Entity/Homeworks.java
@@ -3,10 +3,14 @@ package kireev.ftshw.project.Database.Entity;
 import android.arch.persistence.room.Entity;
 import android.arch.persistence.room.PrimaryKey;
 
+import java.util.List;
+
+import kireev.ftshw.project.Courses.Rating.HomeworkVO;
+
 @Entity
 public class Homeworks {
     @PrimaryKey
-    public long id;
+    public int id;
 
     public String title;
 }
